+++ src/Data/constants.js
@@ -1,169 +1,124 @@
 export const navData = [
-    {
-        title: "Home",
-        path: "/"
-    },
-    {
-        title: "Films",
-        path: "/films",
-    }
+  {
+      title: "Home",
+      path: "/"
+  },
+  {
+      title: "Movies",
+      path: "/movies",
+  },
+  {
+      title: "Sign up",
+      path: "signup"
+  },
+  {
+      title: "Log In",
+      path: "login"
+  },
+  {
+      title: "My Profile",
+      path: "/profile"
+  }
 ]
 
-export const topFilms = [
-    {
-        "id": 1,
-        "title": "The Shawshank Redemption",
-        "year": 1994,
-        "genre": [
-            "Drama"
-        ],
-        "rating": 9.3,
-        "director": "Frank Darabont",
-        "actors": [
-            "Tim Robbins",
-            "Morgan Freeman"
-        ],
-        "plot": "Two imprisoned men bond over several years, finding solace and eventual redemption through acts of common decency.",
-        "poster": "https://m.media-amazon.com/images/M/MV5BNDE3ODcxYzMtY2YzZC00NmNlLWJiNDMtZDViZWM2MzIxZDYwXkEyXkFqcGdeQXVyNjAwNDUxODI@._V1_.jpg",
-        "trailer": "https://example.com/shawshank_redemption_trailer.mp4",
-        "runtime": 142,
-        "awards": "Nominated for 7 Oscars",
-        "country": "USA",
-        "language": "English",
-        "boxOffice": "$28.3 million",
-        "production": "Columbia Pictures",
-        "website": "http://www.warnerbros.com/movies/shawshank-redemption"
-    },
-    {
-        "id": 2,
-        "title": "The Godfather",
-        "year": 1972,
-        "genre": [
-            "Crime",
-            "Drama"
-        ],
-        "rating": 9.2,
-        "director": "Francis Ford Coppola",
-        "actors": [
-            "Marlon Brando",
-            "Al Pacino",
-            "James Caan"
-        ],
-        "plot": "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son.",
-        "poster": "https://th.bing.com/th/id/R.1199dc6273680f175fd9b06c9c36d08a?rik=%2fKp12KVFsHU89w&pid=ImgRaw&r=0",
-        "trailer": "https://example.com/the_godfather_trailer.mp4",
-        "runtime": 175,
-        "awards": "Won 3 Oscars",
-        "country": "USA",
-        "language": "English",
-        "boxOffice": "$245.1 million",
-        "production": "Paramount Pictures",
-        "website": "https://www.paramountmovies.com/movies/the-godfather"
-    },
-    {
-        "id": 3,
-        "title": "The Dark Knight",
-        "year": 2008,
-        "genre": [
-            "Action",
-            "Crime",
-            "Drama"
-        ],
-        "rating": 9,
-        "director": "Christopher Nolan",
-        "actors": [
-            "Christian Bale",
-            "Heath Ledger",
-            "Gary Oldman"
-        ],
-        "plot": "When the menace known as the Joker wreaks havoc and chaos on the people of Gotham, Batman must accept one of the greatest psychological and physical tests of his ability to fight injustice.",
-        "poster": "https://th.bing.com/th/id/R.34df6b287faf65b160b47e05ba85bb3b?rik=o5LsP1up4eGlXA&pid=ImgRaw&r=0",
-        "trailer": "https://example.com/the_dark_knight_trailer.mp4",
-        "runtime": 152,
-        "awards": "Won 2 Oscars",
-        "country": "USA",
-        "language": "English",
-        "boxOffice": "$1.005 billion",
-        "production": "Warner Bros. Pictures",
-        "website": "https://www.warnerbros.com/movies/dark-knight"
-    },
-    {
-        "id": 4,
-        "title": "Pulp Fiction",
-        "year": 1994,
-        "genre": [
-            "Crime",
-            "Drama"
-        ],
-        "rating": 8.9,
-        "director": "Quentin Tarantino",
-        "actors": [
-            "John Travolta",
-            "Uma Thurman",
-            "Samuel L. Jackson"
-        ],
-        "plot": "The lives of two mob hitmen, a boxer, a gangster's wife, and a pair of diner bandits intertwine in four tales of violence and redemption.",
-        "poster": "https://th.bing.com/th/id/OIP.z6bKjfyEmcYCRRPy4b0_KgHaLH?rs=1&pid=ImgDetMain",
-        "trailer": "https://example.com/pulp_fiction_trailer.mp4",
-        "runtime": 154,
-        "awards": "Won 1 Oscar",
-        "country": "USA",
-        "language": "English",
-        "boxOffice": "$213.9 million",
-        "production": "Miramax Films",
-        "website": "https://www.miramax.com/movie/pulp-fiction"
-    },
-    {
-        "id": 5,
-        "title": "Forrest Gump",
-        "year": 1994,
-        "genre": [
-            "Drama",
-            "Romance"
-        ],
-        "rating": 8.8,
-        "director": "Robert Zemeckis",
-        "actors": [
-            "Tom Hanks",
-            "Robin Wright",
-            "Gary Sinise"
-        ],
-        "plot": "The presidencies of Kennedy and Johnson, the events of Vietnam, Watergate, and other historical events unfold through the perspective of an Alabama man with an IQ of 75.",
-        "poster": "https://image.tmdb.org/t/p/original/saHP97rTPS5eLmrLQEcANmKrsFl.jpg",
-        "trailer": "https://example.com/forrest_gump_trailer.mp4",
-        "runtime": 142,
-        "awards": "Won 6 Oscars",
-        "country": "USA",
-        "language": "English",
-        "boxOffice": "$677.9 million",
-        "production": "Paramount Pictures",
-        "website": "https://www.paramountmovies.com/movies/forrest-gump"
-    },
-    {
-        "id": 6,
-        "title": "Inception",
-        "year": 2010,
-        "genre": [
-            "Action",
-            "Adventure",
-            "Sci-Fi"
-        ],
-        "rating": 8.8,
-        "director": "Christopher Nolan",
-        "actors": [
-            "Leonardo DiCaprio",
-            "Joseph Gordon-Levitt",
-            "Ellen Page"
-        ],
-        "plot": "A thief who steals corporate secrets through the use of dream-sharing technology is given the inverse task of planting an idea into the mind of a C.E.O.",
-        "poster": "https://th.bing.com/th/id/R.dd9827fbb0199cd20694739d54c346d8?rik=5JsQjhudqNqO9w&pid=ImgRaw&r=0",
-        "trailer": "https://example.com/inception_trailer.mp4",
-        "runtime": 148,
-        "awards": "Won 4 Oscars",
-        "country": "USA",
-        "language": "English",
-        "boxOffice": "$829.9 million",
-        "production": "Warner Bros. Pictures",
-        "website": "https://www.warnerbros.com/movies/inception"
-    }
+export const navDataLogged = [
+  {
+      title: "Home",
+      path: "/"
+  },
+  {
+      title: "Movies",
+      path: "/movies",
+  },
+  {
+      title: "My Profile",
+      path: "/profile"
+  }
+]
+
+export const tableData = [
+  {
+    title: 'full name',
+    key: 'fullName',
+  },
+  {
+    title: 'Email',
+    key: 'email',
+  },
+  {
+    title: 'username',
+    key: 'username',
+  },
+];
+
+export const datas = [
+  {
+    id: 1,
+    cover:
+      'https://m.media-amazon.com/images/M/MV5BNDE3ODcxYzMtY2YzZC00NmNlLWJiNDMtZDViZWM2MzIxZDYwXkEyXkFqcGdeQXVyNjAwNDUxODI@._V1_.jpg',
+    trailer: 'https://www.youtube.com/embed/PLl99DlL6b4?si=GMKBVj4A-8IquQZ-',
+    title: 'The Shawshank Redemption',
+    year: 1994,
+    genre: ['Drama'],
+    rating: 9.3,
+    director: 'Frank Darabont',
+    awards: 'Nominated for 7 Oscars',
+    country: 'USA',
+    production: 'Columbia Pictures',
+  },
+  {
+    id: 2,
+    cover:
+      'https://m.media-amazon.com/images/M/MV5BM2MyNjYxNmUtYTAwNi00MTYxLWJmNWYtYzZlODY3ZTk3OTFlXkEyXkFqcGdeQXVyNzkwMjQ5NzM@._V1_.jpg',
+    trailer: 'https://www.youtube.com/embed/UaVTIH8mujA?si=IuG1wCLANrY3Ecxb',
+    title: 'The Godfather',
+    year: 1972,
+    genre: ['Crime', 'Drama'],
+    rating: 9.2,
+    director: 'Francis Ford Coppola',
+    awards: 'Won 3 Oscars',
+    country: 'USA',
+    production: 'Paramount Pictures',
+  },
+  {
+    id: 3,
+    cover:
+      'https://upload.wikimedia.org/wikipedia/en/1/1c/The_Dark_Knight_%282008_film%29.jpg',
+    trailer: 'https://www.youtube.com/embed/EXeTwQWrcwY?si=0tYdBXDUMtxE6g5-',
+    title: 'The Dark Knight',
+    year: 2008,
+    genre: ['Action', 'Crime', 'Drama'],
+    rating: 9,
+    director: 'Christopher Nolan',
+    awards: 'Won 2 Oscars',
+    country: 'USA',
+    production: 'Warner Bros. Pictures',
+  },
+  {
+    id: 4,
+    cover: 'https://cdn.europosters.eu/image/1300/1288.jpg',
+    trailer: 'https://www.youtube.com/embed/tGpTpVyI_OQ?si=D4gposHQm2N6U872',
+    title: 'Pulp Fiction',
+    year: 1994,
+    genre: ['Crime', 'Drama'],
+    rating: 8.9,
+    director: 'Quentin Tarantino',
+    awards: 'Won 1 Oscar',
+    country: 'USA',
+    production: 'Miramax Films',
+  },
+  {
+    id: 5,
+    cover:
+      'https://m.media-amazon.com/images/I/61+nkPEbcAL._UF1000,1000_QL80_.jpg',
+    trailer: 'https://www.youtube.com/embed/Mj9IA9tTfio?si=5QtquHPcEfYfzkI4',
+    title: 'Forrest Gump',
+    year: 1994,
+    genre: ['Drama', 'Romance'],
+    rating: 8.8,
+    director: 'Robert Zemeckis',
+    awards: 'Won 6 Oscars',
+    country: 'USA',
+    production: 'Paramount Pictures',
+  },
 ]
