--- src/data/constants.js
@@ -1,124 +1,0 @@
-export const navData = [
-    {
-        title: "Home",
-        path: "/"
-    },
-    {
-        title: "Movies",
-        path: "/movies",
-    },
-    {
-        title: "Sign up",
-        path: "signup"
-    },
-    {
-        title: "Log In",
-        path: "login"
-    },
-    {
-        title: "My Profile",
-        path: "/profile"
-    }
-]
-
-export const navDataLogged = [
-    {
-        title: "Home",
-        path: "/"
-    },
-    {
-        title: "Movies",
-        path: "/movies",
-    },
-    {
-        title: "My Profile",
-        path: "/profile"
-    }
-]
-
-export const datas = [
-    {
-      id: 1,
-      cover:
-        'https://m.media-amazon.com/images/M/MV5BNDE3ODcxYzMtY2YzZC00NmNlLWJiNDMtZDViZWM2MzIxZDYwXkEyXkFqcGdeQXVyNjAwNDUxODI@._V1_.jpg',
-      trailer: 'https://www.youtube.com/embed/PLl99DlL6b4?si=GMKBVj4A-8IquQZ-',
-      title: 'The Shawshank Redemption',
-      year: 1994,
-      genre: ['Drama'],
-      rating: 9.3,
-      director: 'Frank Darabont',
-      awards: 'Nominated for 7 Oscars',
-      country: 'USA',
-      production: 'Columbia Pictures',
-    },
-    {
-      id: 2,
-      cover:
-        'https://m.media-amazon.com/images/M/MV5BM2MyNjYxNmUtYTAwNi00MTYxLWJmNWYtYzZlODY3ZTk3OTFlXkEyXkFqcGdeQXVyNzkwMjQ5NzM@._V1_.jpg',
-      trailer: 'https://www.youtube.com/embed/UaVTIH8mujA?si=IuG1wCLANrY3Ecxb',
-      title: 'The Godfather',
-      year: 1972,
-      genre: ['Crime', 'Drama'],
-      rating: 9.2,
-      director: 'Francis Ford Coppola',
-      awards: 'Won 3 Oscars',
-      country: 'USA',
-      production: 'Paramount Pictures',
-    },
-    {
-      id: 3,
-      cover:
-        'https://upload.wikimedia.org/wikipedia/en/1/1c/The_Dark_Knight_%282008_film%29.jpg',
-      trailer: 'https://www.youtube.com/embed/EXeTwQWrcwY?si=0tYdBXDUMtxE6g5-',
-      title: 'The Dark Knight',
-      year: 2008,
-      genre: ['Action', 'Crime', 'Drama'],
-      rating: 9,
-      director: 'Christopher Nolan',
-      awards: 'Won 2 Oscars',
-      country: 'USA',
-      production: 'Warner Bros. Pictures',
-    },
-    {
-      id: 4,
-      cover: 'https://cdn.europosters.eu/image/1300/1288.jpg',
-      trailer: 'https://www.youtube.com/embed/tGpTpVyI_OQ?si=D4gposHQm2N6U872',
-      title: 'Pulp Fiction',
-      year: 1994,
-      genre: ['Crime', 'Drama'],
-      rating: 8.9,
-      director: 'Quentin Tarantino',
-      awards: 'Won 1 Oscar',
-      country: 'USA',
-      production: 'Miramax Films',
-    },
-    {
-      id: 5,
-      cover:
-        'https://m.media-amazon.com/images/I/61+nkPEbcAL._UF1000,1000_QL80_.jpg',
-      trailer: 'https://www.youtube.com/embed/Mj9IA9tTfio?si=5QtquHPcEfYfzkI4',
-      title: 'Forrest Gump',
-      year: 1994,
-      genre: ['Drama', 'Romance'],
-      rating: 8.8,
-      director: 'Robert Zemeckis',
-      awards: 'Won 6 Oscars',
-      country: 'USA',
-      production: 'Paramount Pictures',
-    },
-  ];
-
-  export const tableData = [
-    {
-      title: 'full name',
-      key: 'fullName',
-    },
-    {
-      title: 'Email',
-      key: 'email',
-    },
-    {
-      title: 'username',
-      key: 'username',
-    },
-  ];
