+++ src/store/selectors/userInformation.jsx
@@ -1,0 +1,7 @@
+const userSelector = (state) => state.profile.loggedUser;
+const authSelector = (state) => state.auth.auth.isAuth;
+
+export {
+   userSelector,
+   authSelector 
+  };
